Se arregla todo lo de movimientos de venta

+++ ezra-backend-spring-boot-apirest/src/main/java/com/ezra/programandojuntos/models/entity/EstadoPedido.java
@@ -16,16 +16,18 @@ public class EstadoPedido implements Serializable{
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Byte id;
+	private Long id;
 	
 	private String estado;
 	private String descripcion;
 	
-	public Byte getId() {
+
+
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Byte id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -33,25 +35,17 @@ public class EstadoPedido implements Serializable{
 		return estado;
 	}
 
-
-
 	public void setEstado(String estado) {
 		this.estado = estado;
 	}
-
-
 
 	public String getDescripcion() {
 		return descripcion;
 	}
 
-
-
 	public void setDescripcion(String descripcion) {
 		this.descripcion = descripcion;
 	}
-
-
 
 	private static final long serialVersionUID = 1L;
